refactor: common directory 구조 변경 (#48, #49) - filter, interceptor, module 디렉토리 생성 및 파일 이동
