+++ src/api/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 
-import { TypeOrmExModule } from '@/common/typeOrmEx.module';
+import { TypeOrmExModule } from '@/common/modules';
 import { ENV_PROVIDER } from '@/utils/constants';
 
 import { AuthController } from './auth.controller';
