+++ src/api/user/user.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { UserRepository } from '@/api/user/user.repository';
-import { TypeOrmExModule } from '@/common/typeOrmEx.module';
+import { TypeOrmExModule } from '@/common/modules';
 
 import { UserController } from './user.controller';
 import { UserService } from './user.service';
