+++ src/common/modules/index.ts
@@ -1,0 +1,1 @@
+export * from './typeorm-ex.module';
