+++ test/api/auth/auth.controller.spec.ts
@@ -4,8 +4,8 @@ import * as testRequest from 'supertest';
 import { AuthController } from '@/api/auth/auth.controller';
 import { AuthService } from '@/api/auth/auth.service';
 import { GoogleAuthGuard } from '@/api/auth/guards/oauth.guard';
-import { HttpExceptionFilter } from '@/common/http-exception.filter';
-import { SuccessInterceptor } from '@/common/success.interceptor';
+import { HttpExceptionFilter } from '@/common/filters';
+import { SuccessInterceptor } from '@/common/interceptors';
 import { USER_STUB } from 'test/api/user/mock';
 import { createMockAuthGuard } from 'test/utils/createMockJwtAuthGuard';
 import createTestingModule from 'test/utils/createTestingModule';
